% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_from_inbox.R
\name{remove_from_inbox}
\alias{remove_from_inbox}
\title{Title}
\usage{
remove_from_inbox(messages_id = "Label_1", remove_labels = "INBOX")
}
\arguments{
\item{messages_id}{}
}
\description{
Title
}
