% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_seafile.R
\name{upload_to_seafile}
\alias{upload_to_seafile}
\title{upload_to_seafile}
\usage{
upload_to_seafile(files, output_directory = folder_upload(),
  repos_id_destination = get_repo_id())
}
\arguments{
\item{files}{files to upload}

\item{repos_id_destination}{repos id}

\item{dossier_upload}{path in the repos}
}
\description{
upload_to_seafile
}
